refactor!: separate file for constants and custom hooks
