+++ react/project-tictac/src/App.jsx
@@ -1,42 +1,23 @@
 /* eslint-disable react/prop-types */
 
-/**
- * - [x] for the current move only, show 'you are at move # ' instead of a button.
- * - [x] rewrite `Board` to use two loops to make the squares instead of hardcoding them.
- * - [x] add a toggle button that lets you sort the moves in either ascending or descending order.
- * - [x] when someone wins, highlight the three squares that caused the win.
- * - [x] when no one wins, display a message about the result being a draw.
- * - [x] display the location for each move in the format (row, col) in the move history list.
- */
+/* improvement: separate file for constants and hooks, and remove magic number */
 
-import { useState } from 'react';
+import { BOARD_SIZE, WINNING_LINES } from './constants';
+import { useGame } from './hooks/useGame';
 
 function Game() {
-  // history initiated with an array that contains a single item, which itself is an array of 9 `null`.
-  // each item will snapshot the square in each move
-  const [history, setHistory] = useState([Array(9).fill(null)]);
-  const [record, setRecord] = useState([[0, 0]]);
-  const [currentMove, setCurrentMove] = useState(0);
-  const [isAscending, setIsAscending] = useState(true);
-
-  const xIsNext = currentMove % 2 === 0;
-  const isEnd = currentMove === 9;
-  const currentSquares = history[currentMove];
-
-  function handleSort() {
-    setIsAscending(!isAscending);
-  }
-
-  function handlePlay(nextSquares, i) {
-    const row = Math.floor(i / 3) + 1;
-    const col = (i % 3) + 1;
-    const nextHistory = [...history.slice(0, currentMove + 1), nextSquares];
-    const nextRecord = [...record.slice(0, currentMove + 1), [row, col]];
-
-    setHistory(nextHistory);
-    setRecord(nextRecord);
-    setCurrentMove(nextHistory.length - 1);
-  }
+  const {
+    history,
+    record,
+    currentMove,
+    isAscending,
+    xIsNext,
+    isEnd,
+    currentSquares,
+    handleSort,
+    handlePlay,
+    setCurrentMove,
+  } = useGame();
 
   return (
     <div className='game'>
@@ -98,9 +79,9 @@ function Board({ xIsNext, isEnd, squares, onPlay }) {
         {!calResult(squares) && !isEnd && `Next: ${xIsNext ? 'X' : 'O'}`}
         {!calResult(squares) && isEnd && `Game draw`}
       </div>
-      {[0, 1, 2].map((row) => (
+      {[...Array(BOARD_SIZE)].map((_, row) => (
         <div key={row} className='board-row'>
-          {[0, 1, 2].map((col) => {
+          {[...Array(BOARD_SIZE)].map((_, col) => {
             // calculate the square index. converts 2D coordinates into 1D array index
             const index = row * 3 + col;
             return (
@@ -134,20 +115,9 @@ function App() {
 }
 
 function calResult(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
+  for (const [a, b, c] of WINNING_LINES) {
     if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return { winner: squares[a], line: lines[i] };
+      return { winner: squares[a], line: [a, b, c] };
     }
   }
   return null;
